Reorganize the tests into helpful suites.
