+++ src/test/java/net/tullco/addressbook/test_classes/AddressTest.java
@@ -1,0 +1,44 @@
+package net.tullco.addressbook.test_classes;
+
+import static org.junit.Assert.*;
+
+import org.junit.Before;
+import org.junit.BeforeClass;
+import org.junit.Test;
+
+public class AddressTest {
+
+	@BeforeClass
+	public static void setUpBeforeClass() throws Exception {
+	}
+
+	@Before
+	public void setUp() throws Exception {
+	}
+
+	@Test
+	public void testSave() {
+		fail("Not yet implemented");
+	}
+
+	@Test
+	public void testDelete() {
+		fail("Not yet implemented");
+	}
+
+	@Test
+	public void testAddressLoader() {
+		fail("Not yet implemented");
+	}
+
+	@Test
+	public void testAddressesLoader() {
+		fail("Not yet implemented");
+	}
+
+	@Test
+	public void testAddAddressLink() {
+		fail("Not yet implemented");
+	}
+
+}
