+++ src/test/java/net/tullco/addressbook/test_suites/AllTests.java
@@ -1,0 +1,11 @@
+package net.tullco.addressbook.test_suites;
+
+import org.junit.runner.RunWith;
+import org.junit.runners.Suite;
+import org.junit.runners.Suite.SuiteClasses;
+
+@RunWith(Suite.class)
+@SuiteClasses({ ModelTests.class })
+public class AllTests {
+
+}
